Move navigation and navigable content to common source set
